+++ scripts/storage.js
@@ -1,6 +1,6 @@
 const safe = require(p.join(appDir, 'scripts/storage/safe.js'));
 
-const changes = 118; // Update this if readingPagesConfig is updated
+const changes = 119; // Update this if readingPagesConfig is updated
 
 const readingPagesConfig = {
 	readingConfigName: '',